backend/requirements.txt
# Google ADK and AI
google-adk>=0.1.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0

# Google Cloud Services
google-cloud-bigquery>=3.14.0
firebase-admin>=6.4.0
google-auth>=2.27.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
